+++ packages/camera/camera/lib/src/camera_controller.dart
@@ -235,13 +235,7 @@ class CameraController extends ValueNotifier<CameraValue> {
     await CameraPlatform.instance.prepareForVideoRecording();
   }
 
-  /// Captures an image and saves it to [path].
-  ///
-  /// A path can for example be obtained using
-  /// [path_provider](https://pub.dartlang.org/packages/path_provider).
-  ///
-  /// If a file already exists at the provided path an error will be thrown.
-  /// The file can be read as this function returns.
+  /// Captures an image and returns the file where it was saved.
   ///
   /// Throws a [CameraException] if the capture fails.
   Future<XFile> takePicture() async {