+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.3
+
+* Adds torch mode as a flash mode for Android and iOS implementations.
+
+## 0.6.2+1
+
+* Fix the API documentation for the `CameraController.takePicture` method.
+
 ## 0.6.2
 
 * Add zoom support for Android and iOS implementations.